Remove conversion to ks in favor of syncronization only. Move synchronization into the main pipeline
